bug_reporter_agent/requirements.txt
Flask==2.3.2
google-cloud-pubsub==2.21.0
google-cloud-firestore==2.11.1
